Move test_posdef_sym into test_multivariate
