+++ .github/workflows/tests.yml
@@ -41,7 +41,6 @@ jobs:
             pymc/tests/distributions/test_logprob.py
             pymc/tests/test_aesaraf.py
             pymc/tests/test_math.py
-            pymc/tests/test_posdef_sym.py
             pymc/tests/backends/test_ndarray.py
             pymc/tests/test_hmc.py
             pymc/tests/test_func_utils.py
